auth: type alias for initial response and session
Unfortunately, starting from the following commit we will be boxing
the futures returned by `AuthenticatorProvider`.

This requires to wrap the returned type with `BoxFuture`, and so the
resulting type becomes complex which results in clippy complaints.

This is why we introduce a type alias for the pair representing an
initial auth response and a boxed auth session.