metrics: remove Histogram from MetricsError
Public MetricsError enum contains a reference to `Histogram`.

In fact, the `MetricsError::Poison` is never really constructed
since we always unwrap the PoisonError when trying to acquire
the mutex.

This is why we get rid of the `MetricsError::Poison` enum value and
convert the `MetricsError` enum to struct holding an error cause.