semver_checks CI: Edit posted comment when possible
Up until now semver_checks CI posted a new comment after each push to
a PR if such push resulted in PR (still) having breaking changes.
This was quite spammy for PRs that intentionally made breaking changes,
because each and every push resulted in a new comment.

This commit changes the CI so that it edits already posted comment,
and only posts a new one when there was none before.

If there are multiple comments already (as is the case with already
open PRs) then the oldest one will be chosen for edit.
The aim is to always have semver status immediately below PR description.