result: decouple TableSpec and ColumnSpec
Every column in a Response::Result comes from the same table (and it
follows that from the same keyspace as well). It's thus redundant
to store a copy of TableSpec (table and keyspace names) in each
ColumnSpec (name and type of a column), which was done in the code
before.
Correctness of the above assumption is checked: if any two columns
of the same Result differ in keyspace or table names, ParseError is
returned.

This commit moves TableSpec out of ColumnSpec and only allocates it once
per query response, effectively saving `2 * (C-1)` string allocations,
where `C` denotes the number of columns returned in the response.
TableSpec is now stored in ResultMetadata and PreparedMetadata.
A getter for table_spec in RowIterator is provided for users and for
existing tests.