Treewide: Remove unreachable `pub`
Some structs / fields etc were marked as pub despite not being a part of
public API. This is confusing when reading code: in order to know if
something is truly public you need to check all the modules up to lib.rs
to verify that each appears in some `pub use`.

This commit removes all such occurrences - now something is public if and
only if it is marked as pub.