scylla unit tests: introduce setup_tracing()
Before, tracing crate's prints weren't captured during tests. Those with
error level used to bloat tests' output, which was especially painful in
CI. OTOH, we wanted to have traces visible in failed tests in order to
easier spot the cause of a failure.
By using tracing_subscriber's TestWriter, tracing's prints are now
captured and only shown if a test fails. This enables efficient
troubleshooting especially in case of flaky tests, as those are common
in CI only and hard to be reproduced locally.
For all tests to easily use the captured tracing, a function
setup_tracing() is introduced. This commit adds it to scylla crate unit
tests. Integration tests and proxy tests are covered in next commits.