session_test: Increase tracing info timeout
Default values for fetching tracing info are quite strict in the driver:
it will make 5 attempts to fetch the data, waiting 3ms between attempts.

For Scylla this is enough, but for Cassandra it is not. For this reason
our tests sometimes fail.

This was partially (for one test) fixed previously:
https://github.com/scylladb/scylla-rust-driver/commit/61ad8781f1a01c9ec5ff087296229c2245c429b3

This commit fixes rest of the tests in session_test that could be affected.
It also changes timeout parameters. Previous fixed used 50 attempts with
200ms wait, I changed it to 200 attempts with 50ms wait in order to not
slow down Scylla tests unnecessarily.