deserialize: introduce FrameSlice
FrameSlice is going to be the main hero of the new lazy deserialization
framework. FrameSlice, unsurprisingly, represents a slice of a result
frame. What is interesting though is that apart from a regular &[u8]
slice, it holds a Bytes object that keeps the whole frame alive. This
allows deserializing blobs as owned ('static!) objects, by subslicing
the original Bytes object.
This commit contains a suite of tests for FrameSlice.

Co-authored-by: Piotr Dulikowski <piodul@scylladb.com>