Token: add constructor and normalization
`i64::MIN` turns out to be an invalid value for a token:
https://github.com/scylladb/scylladb/blob/4be70bfc2bc7f133cab492b4aac7bab9c790a48c/dht/token.hh#L32
Hashing algorithm should change `i64::MIN` result to `i64::MAX`.
In order to have this check in one place and make sure we always perform
it I made `value` field of `Token` non-public and created constructor and
getter for it.