CI, Makefile: clippy and check with all-features
In the previous version of tablet PR tablets were put behind
"unstable-tablets" feature. In order to test them the feature had to
be enabled, but CI / Makefile didn't do this for cargo check or
clippy. While the tablets are no longer hidden behind feature flag,
I think checking with all features is still a valid addition, so I
decided to leave this commit as-is.