types: return std::io::Error instead of ParseError
It makes no sense for the functions to return broader error type while
only a proper subset of it can be returned. In the particular case,
std::io::Error is needlessly converted to ParseError, which makes
pattern matching required in the caller in order to get the
std::io::Error back.
The functions are changed to return std::io::Error. The previous
behaviour of propagating the error as ParseError can be easily
achieved using ? operator in the caller:

```rust
fn i_return_parse_error(buf: &mut &[u8]) -> Result<(), ParseError) {
	read_int(buf)?;
	Ok(())
}
```