scylla-cql/types: introduce new deserialization traits
This commit introduces traits that will serve as the base of the new
deserialization framework. It is meant to replace the old way of
deserialization based on CqlValue, FromCqlVal and FromRow.

The new traits are `DeserializeValue<'f>` and `DeserializeRow<'f>`. If a
type implements `DeserializeValue<'f>`, this means that it can take a
buffer of lifetime `'f` and deserialize itself, treating the contents of
the buffer as a CQL value. Analogously, `DeserializeRow<'f>` allows for
deserialization of types that are supposed to represent a whole row
returned in the results.

Deserialization is now split into two phases: type checking and actual
deserialization. Both traits have two methods: `type_check` and
`deserialize`. The idea behind this is to validate the column types of
the response only once, after the response is received. The
`deserialize` method is then called for each row/value to perform the
actual parsing: that method is allowed to assume that `type_check` was
called and may skip some type checking for performance (although it
cannot use that assumption to perform unsafe operations, it is not an
unsafe method).

The previous framework only supported deserialization of types that
exclusively own their data, for example String or Vec<u8>. However, the
new framework allows for types that borrow or co-own the frame buffer:

- Types that borrow the buffer are &'f str and &'f [u8]. They just point
to the raw data from the serialized response. Rust's lifetime system
makes sure that the the user doesn't deallocate the serialized
response until using the deserialized types that point to the buffer
are dropped.
Apart from the aforementioned types, a bunch of iterator types are
introduced that allow consuming collections without allocations.
- Serialized frame is represented with bytes::Bytes. It is possible to
create a subslice (also of type Bytes) that keep the original Bytes
alive. The type being deserialized can obtain access to such a
subslice. Like in the case of e.g. &str vs. String, keeping Bytes also
allows to avoid an allocation, but it is also easier to handle as the
deserialized type doesn't have to be bound by any lifetime. It is
important to be careful when handling such types as they keep the
whole serialized frame alive, even if they only point to a small
subslice. This can lead to a space leak and more memory being used
than necessary.

Co-authored-by: Piotr Dulikowski <piodul@scylladb.com>