cargo: put restrictions on tokio version
We decided to bump minimum required version of tokio to 1.34.

Currently, the newest tokio version is 1.38, but some of the integration
tests are eaten when testing with this specific verstion of tokio.
Which is why, as of now, we decided not to support this version.

The issue with version 1.38 is related
to #[tokio::test] and #[ntest::timeout] attributes.
Refs:
- https://github.com/tokio-rs/tokio/issues/6610
- https://github.com/becheran/ntest/pull/28
- https://github.com/tokio-rs/tokio/pull/6497