f_errors: derive Clone
In the following commit we want to introduce a
CqlResponseParseError type which will be a new variant
of `[Query/NewSession]Error`. These two error types
implement a `Clone` trait.

This is why, we need to derive Clone for all of the types
introduces in previous commits.

Unfortunately, the `std::io::Error` does not implement Clone.
This is why we need to wrap any appearing `std::io::Error` with
`Arc`, so the enum types encapsulating it can derive Clone.