frame/startup: store options as Cow<'a, str>
Majority of, if not all, keys sent in the string map in STARTUP message
by the driver are known in the compile time. Some values may not be
known at compile time, but still it makes more sense to borrow them from
a String rather than clone each time a connection is opened. To reduce
needless allocations while preserving ability to add keys and values
dynamically, the HashMap type is changed to
`HashMap<Cow<'a, str>, Cow<'a, str>>`.
This is going to reduce number of allocations upon opening each
connection.