connection: resolve a dated TODO
Answering the question there: no, that logic shouldn't be in
`Connection::new`. `Connection::new`'s responsibility is to create a
connection and make it ready to send CQL frames. OTOH,
`open_named_connection` is responsible for connection setup on CQL
level, i.e. for OPTIONS/STARTUP handshake and registering for events
(if this is a control connection).

Informative docstrings are added that convey the above separation.