conn: copy a batch type in prepare_batch
Previously, there was a bug in prepare_batch function.
We wouldn't inherit a BatchType from the `initial_batch` - in result,
we would always send LOGGED batches (since LOGGED is a default batch type).

This would result in server errors when including a counter statements
in a LOGGED batch. I added a test case which checks that it works
correctly now, and the driver sends a COUNTER batch when user asks to.