topology: skip invalid peers instead of failing
It appears that in some clusters there sometimes emerges a transient
state that there are some incomplete/invalid peer entries in system
tables. Under such circumstances, the driver would throw an error and
fail the metadata refresh altogether, eventually leading to critical
conditions such as no hosts to query.
To fix this in line with how other drivers approach this, a warning is
emitted upon invalid peer entry read, and the entry is skipped with no
error returned.