requests: make paging state strongly typed
As an effort of more robust paging API, paging state is now made typed.
Two new types are introduced: PagingState and PagingStateResponse.
PagingState is, underneath, just an Arc over the raw bytes from
the DB. The idea is that PagingState is passed as an argument to
request execution API to ask for resuming the query (_continue it_) from
the given PagingState. PagingStateResponse is returned always by the DB
from paged requests - with information about either more or no more
pages available. PagingState can be easily retrieved from
PagingStateResponse (if it contains some), so that user can use it in
a loop.
The select_paging example is modernised to showcase the idea.