result: introduce RawRowLendingIterator
The iterator is analogous to RawRowIterator, but instead of borrowing
from external frame bytes and metadata with 'frame lifetime, it owns
them and lends them from itself. Thus, it cannot implement Iterator
trait. It does not, however, prevent us from exposing a `next()` method
on it.

The iterator is going to be used in new iterator API for queries
(i.e., the one connected to `{query,execute}_iter`), where borrowing is
not suitable (or even possible) due to the design of that API.

Tests of RowIterator are shared with the new RawRowLendingIterator,
by introducing a new LendingIterator trait using GATs.
Due to a bug/limitation in the compiler, 'static lifetimes are needed in
tests. I'd like to use the recently stabilised (1.80)
std::sync::LazyLock, but our MSRV is too old. Instead, I've employed
lazy_static.