session: {query,execute}_single_page pass page state explicitly
As an attempt to make paged queries' API more explicit, robust and
self-explanatory, PagingState is decoupled from QueryResult. Instead,
it's returned from {query,execute}_single_page as a second field of a
pair, helping users not to forget about it.
Moreover, internal and external APIs which expect that the query is not
paged now issue an error tracing message and return ProtocolError error
if a "more pages" page state is returned from the unpaged query.