errors: BrokenConnectionError
Introduced a BrokenConnectionError. Since users probably won't really
be interested in enum-matching the cause of this error, we hide
the kind of broken connection error behind Arc<dyn Error>.

Other reason for that is there will be a cycle of dependency between
BrokenConnectionError and RequestError types
(once RequestError is introduced later in this PR). As of now, the cycle
is also visible, because of QueryError.
We don't want to expose a cyclic error types to the users.

The cause of the cycle is:
- Request can fail due to connection being broken (RequestError -> BrokenConnectionError)
- A keepaliver sends a request, which can fail. If keepaliver, returns an
  error, we close the connection. The cause of broken connection is now a
  keepaliver's request error (BrokenConnection -> RequestError)