connection: introduce ConnectionError
Yet another error type introduced. This time, as name suggests,
we introduce an error type corresponding to errors that happened
around a specific connection.

Notice, that this error type will not be accessed by the user.
It will be handled by PoolRefiller. Each error will obviously be logged,
but this logic is already implemented in PoolRefiller.

This error type is introduced to narrow the type of errors handled
by PoolRefiller (which happen on connection-level).

We also implement a `is_address_available_for_use` method
which is used during source port selection for a given shard.
In a following commit, this method will be removed from `QueryError`.