errors: CqlEventHandlingError
Introduced an error returned by `handle_event` method.
In addition, adjusted the `handle_event` method's logic, so we
propagate the error in case of receiving invalid response (other than EVENT)
on stream -1. This way, we assume that this connection should be closed,
and so the error will be handled by pool refiller.