errors: introduce UserRequestError
This is an error that appears case of
failure of user request (requests triggered via public API, i.e.
PREPARE, EXECUTE, QUERY, BATCH).

Its only purpose is to filter out impossible error variants from
`CqlResponseParseError` that appear in `RequestError` type. For example,
it should not be possible to receive `CqlSupportedParseError` when parsing
the response to user request. It means that the server sent an invalid
response to the corresponding request.

The From<UserRequestError> implementation will be adjusted in a following
commits. Firstly, we need to narrow the type of corresponding functions
and methods to UserRequestError.