result: extract paging_state from ResultMetadata
Although paging state comes from the DB in the <result_metadata> CQL
protocol entity, it should be considered separate. The main reason for
that is that PreparedStatement also receives (and optionally caches)
ResultMetadata upon preparation, which is then used for an optimisation
(col_specs are not sent in RESULT::Rows response by the DB). However,
paging state should not be cached this way, as it's going to change
in between execution of the same statement with different paging state
given.
In the next commit ResultMetadata is going to be shared between
PreparedStatement and its QueryResults by an Arc, and for that paging
state must be kept separate.