connection: query_single_page takes paging state
It appears that query_single_page's previous semantics was to only fetch
the first page and ignore others. It makes much more sense, though, to
have its semantics consistent with Session::query_single_page: to take
PagingState and query a single page pointed by that state. This allows
using it in a loop, similarly to Session::query_single_page.