scylla: introduce errors module
Currently, most of the error types are contained in scylla-cql crate.
This makes it much harder to make some errors pub(crate) (re-exporting
is required). In addition, some error types, such as `QueryError`
and `NewSessionError` should not belong to scylla-cql crate. They
belong to public API and should be defined in `scylla` crate.