response: remove UserRequestError from Response::into_non_error_response
We want to move `UserRequestError` to scylla crate.
To achieve this, we cannot have any places in `scylla-cql` crate
that would depend on this error's definition. Fortunately, there
was only one place that needed it. To fix it, we simply narrow
the return type of the method and introduce a From<> implementation
for the transition. This way, we do not have to adjust the
callers (that use `?` operator).