errors: move BrokenConnectionErrorKind to scylla crate
NOTE: in this commit, we only move BCErrorKind (without BCError), since
there is a cyclic dependency BCError -> Arc<BCErrorKind> -> RequestError -> ...
BCError will be moved in a later commit.

We also temporarily make the only field of BCError public, so we can
create it from BCErrorKind in From<> implementation (residing in scylla crate).