errors: remove scylla-cql::errors module
The only error type (and its dependencies) that was left in this module
was DbError type. It directly corresponds to `response::error::Error` type.
I think the `response::error` module is a great place to put it there.

We re-export this type in scylla crate, so it can be found in the same
place as other error types (scylla::errors, previously scylla::transport::errors).

We can also now remove the scylla-cql::errors module altogether.