docs-pages.yml: Trigger on tag push instead of branch-* push
This workflow should be triggered when the generated output may be
different than before.
This can happen when `main` branch is updated - because it contains
configuration of the documentation (selects which tags are rendered,
which ones are deprecated etc).
It can also happen when a tag is pushed:
- it may have been moved to a different commit so the content will be
  different.
- it may be a new tag that was previously listed in conf.py on main branch,
  and so was unavailable during generation.

Pushing to `branch-*` can't change generated output in any way - so this
commit removes this trigger.

It also adds trigger on tag push - for the reasons listed above.