errors: expose downcast_ref for type-erased errors
Sometimes, user might want to downcast `Arc<dyn Error>` to specific
error type.

Actually, there is a use case for this in cpp-rust-driver,
where we need to define custom `SerializeRow` implementation. It returns
a custom error type. We would like to match against it during
rust-to-C error conversion - downcasting is thus required.