scylla/lib.rs: Remove stars in serialization frameworks imports
It is easier to see what is imported, and decide if it should be,
if one can see what exactly is imported. This is hard to do with stars,
so, as part of API stabilization effort, those are removed here.

For now, I did not remove any imports, so this change should not
be breaking.