cloud: Export `CloudConfig`.
This struct is a part of public API: it is present in a field of
`SessionConfig`. It was not exported - so it was an unnameable type.

It has a pub method that allows its construction by parsing a yaml file,
so user could (if this struct was public) create it and put it in several
`SessionConfigs`.

This commit exports this struct so it is a proper part of public API.