ClusterData: remove get_datacenters_info()
This method seems to have very limited usability:
- For iterating trough datacenter names user can call
  `ReplicaLocator::datacenter_names, which is constant time.
- Nothing else was available to the user because `Datacenter` struct is
  not exported (so is unnameable). This also means all the work done by
  this method was a waste.
- For more advanced tasks there are `ReplicaLocator::unique_nodes_in_datacenter_ring`
  and `ClusterData::get_nodes_info`. I don't think it is worth it to have
  this method which allocates a lot on each call and calculates rack counts.
  It could make sense to have rack counts precalculated when creating
  `ClusterData` - I'll think about it during metadata refactor.

I'm pretty sure that the original intention was to calculate this whole
hash map in `ClusterData::new`, but for some reason the calculation there
was thrown out and redone here. I don't think its worth it repair this
API now considering it was basically unavailable to users.