session_test: Don't use ClusterData::get_datacenters_info
This method does some things that are unnecessary here:
- For each DC it allocates a vector with it's nodes
- It calculates rack count (which requires .unique() call which allocates)

For the purpose of this test - and I'm pretty sure most of possible
usages of get_datacenters_info - it is enough to use `.get_nodes_info()`
and iterate on that.

We don't avoid `unique()` call but we avoid creating vectors of nodes.