deser macros: prepare for second lifetime parameter
As DeserializeValue and DeserializeRow are going to be parametrized by
two lifetimes instead of one, the common code for generating unique
lifetimes in derive macros is adjusted to generate two lifetimes.

For now, only the first generated lifetime is used (the second is
ignored), because DeserializeValue and DeserializeRow still have only
one lifetime.