parametrize DeserializeValue with two lifetimes
`DeserializeValue` requires two types of data in order to perform
deserialization:
1) a reference to the CQL frame (a FrameSlice),
2) the type of the column being deserialized, being part of the
   ResultMetadata.

When deserializing owned types, both the frame and the metadata can have
any lifetime and it's not important. When deserializing borrowed types,
however, they borrow from the frame, so their lifetime must necessarily
be bound by the lifetime of the frame. Metadata is only needed for the
deserialization, so its lifetime does not abstractly bound the
deserialized value.

Up to this commit, DeserializeValue was only parametrized by one
lifetime: 'frame, which bounded both the frame slice and the metadata.
This was done that way due to an assumption that both the metadata and
the frame (shared using Bytes) will be owned by the same entity.
However, the new idea of deserializing result metadata to a borrowed
form (to save allocations) makes result metadata's lifetime shorter than
the frame's lifetime. Not to unnecessarily shorten the deserialized
values' lifetime, a separate lifetime parameter is introduced for
result metadata: 'metadata.

The commit is large, but the changes are mostly mechanical, by adding
the second lifetime parameter.

DeserializeRow & friends are going to get the second lifetime parameter,
too, but for now they pass 'frame as both lifetime parameters of
DeserializeValue.