parametrize DeserializeValue with two lifetimes
`DeserializeValue` requires two types of data in order to perform
deserialization:
1) a reference to the CQL frame (a FrameSlice),
2) the type of the column being deserialized, being part of the
   ResultMetadata.

It's important to understand what is a _deserialized value_. It's not
just an implementor of DeserializeValue; there are some implementors of
`DeserializeValue` who are not yet final values, but **partially**
deserialized types that support further deserialization - _value
deserializers_, such as `ListlikeIterator` or `UdtIterator`.
_Value deserializers_, as they still need to deserialize some value, are
naturally bound by 'metadata lifetime. However, _values_ are completely
deserialized, so they should not be bound by 'metadata - only by 'frame.

When deserializing owned values, both the frame and the metadata can
have any lifetime and it's not important. When deserializing borrowing
values, however, they borrow from the frame, so their lifetime must
necessarily be bound by the lifetime of the frame. Metadata is only
needed for the deserialization, so its lifetime does not abstractly
bound the deserialized value.

Up to this commit, DeserializeValue was only parametrized by one
lifetime: 'frame, which bounded both the frame slice and the metadata.
(why? see next paragraph about value deserializers)
This was done that way due to an assumption that both the metadata and
the frame (shared using Bytes) will be owned by the same entity.
However, the new idea of deserializing result metadata to a borrowed
form (to save allocations) makes result metadata's lifetime shorter than
the frame's lifetime. Not to unnecessarily shorten the deserialized
values' lifetime, a separate lifetime parameter is introduced for
result metadata: 'metadata.

Up to this commit, value deserializers could only be parametrized by
'frame lifetime. Therefore, borrowed metadata held by value
deserializers limits the 'frame lifetime to its lifetime. Hence, values
deserialized by value deserializers must have been bounded by the
lifetime of the metadata.

The commit is large, but the changes are mostly mechanical, by adding
the second lifetime parameter.

DeserializeRow & friends are going to get the second lifetime parameter,
too, but for now they pass 'frame as both lifetime parameters of
DeserializeValue.