Unpub UntranslatedEndpoint and ResolvedContactPoint
`ResolvedContactPoint` was unnameable. It was part of public API,
because it was in a variant of `UntranslatedEndpoint`, which itself was
public.

`UntranslatedEndpoint` is not used in any public API, so can be made
`pub(crate)`. Thanks to that `ResolvedContactPoint` can also be made
`pub(crate)` as it no longer appears in any public API.