result: fix lifetimes in ColumnSpec
When a lifetime is not given explicitly in a value returned from
a method, the rules of lifetime elision make the value bounded by the
lifetime of Self (the implied 's lifetime in `&'s self`).
In case of `ResultMetadata::col_specs`, `ColumnSpec`'s lifetime
parameter was unnecessarily bound to the lifetime of `ResultMetadata`.

Also, the commit lifts the requirement of `ResultMetadata::new_for_test`
for the `ColumnSpec` given as an argument to have `'static` lifetime.