f_errors: RawRowsAndPagingStateResponseParseError
An error that occurred during initial deserialization of
`RESULT:Rows` response. Since the deserialization of rows is lazy,
we initially only need to deserialize:
- result metadata flags
- column count (result metadata)
- paging state response

We ultimately want to get rid of current `RowsParseError` since its usage
is abused in a lot of places throughout the scylla crate. We start by narrowing
the return type of `deser_rows()` function.