metadata: return ResultMetadataAndRowsCountParseError in deserialize_metadata
Adjusted the error returned by `RawMetadataAndRawRows::deserialize_metadata()`.

Again, to narrow the return error type of following methods:
- RawMetadataAndRawRows::deserialize_metadata()
- QueryResult::into_rows_result()

Adjusted the callers of `QueryResult::into_rows_result` in topology.rs and session.rs
and added corresponding variants for `TracingProtocolError` and `SchemaVersionFetchError`

Temporarily, we need to introduce a corresponding variant to `QueryError`
and `NewSessionError`. This is because internal API makes use of `into_rows_result()`
in multiple places. The callers will be adjusted later in this PR, allowing us
to remove the variant from `Query/NSError`.