errors: NextPageError and NextRowError
These are errors returned by async iterator API.

Ultimately, we want them to be independent types returned by the public API.
However, as of now, NextRowError needs to be wrapped in QueryError - to
address that, further changes to iter API need to applied. These, however,
are not in a scope of this PR.

This partial change of iter API is introduced because we want to get rid
of RowsParseError dependency in this module.