macros: SerializeValue supports #[allow_missing]
The `#[allow_missing]` attibute is intented to allow a flexible
transition period when schema is altered. Namely, if a UDT is extended
with a new field and the transition begins with migrating clients to the
new schema (extending the Rust struct with a new field), then server may
still have the old schema and provide metadata that is missing the new
field. In such case, `#[allow_missing]` can be attached to the new Rust
struct's field and this way handle the situation:
- in deserialization, the field missing from DB metadata will be
  default-initialized,
- in serialization, such a field will be ignored and not serialized.

This commit adds support for this attribute: for parsing it using
`darling`, and for its semantics in both `match_by_name` and
`enforce_order` flavors.

Also, corresponding (doc)tests for attribute sanity verification are
added, fully based on those from DeserializeValue.